requirements.txt
requests
termcolor
discord.py
pyourls3